refactor: 새로운 localProducts를 set하는 로직 커스텀훅으로 분리
