+++ src/components/index.ts
@@ -8,6 +8,7 @@ export { GuideBox } from "./GuideBox";
 export { Toast } from "./Toast";
 export { ErrorBoundary } from "./ErrorBoundary";
 
+export { ServerSelectBox } from "./ServerSelectBox";
 export { ProductList } from "./ProductList";
 export { CartProductList } from "./CartProductList";
 export { OrderProductList } from "./OrderProductList";
